Handle errors regarding temperature threshold
If no custom threshold is set we need something to compare to.
If the custom threshold is not a valid value we need to catch ValueError.
If the custom threshold is lower than zero we need to set some value, too.
For all these cases, we use self.THRESHOLD_TEMP as value for temp_thresh.

This fixes issue #87.